Last normal commit before big changes
This commit doesn't have that much changes, but I'm just doing it
because, as you can see, the project is going to get a big reestructure.
The mainting of the code is becoming almost impossible, as everything is
in one single file. In the next commits, I will transfer all parts of
codes to separate folders and files. I have started with this commit,
but I have not implemented anything (as of now). Anyway, here is the
changelog for this version:
- Centralized some of the code about transitions into one singular
  section
- Fixed search pop-up having a bad implementation of a idea I had. Now
  should also be more responsive with other screen resolutions
  - Also changed the layout of the buttons in the search pop-up
- Did a little overhaul in the inbox pop-up, as now it follows other
  buttons and roundness. I will probably work on it for some time as I'm
  not that satifisfied with it right now
- Removed a lot of commented code that will not be used