Another hotfix update
Nothing that big has been added to this update, as it was mainly a
hotfix one. Now the project contains a experimental branch, where I will
add new features that may break stuff. This will ensure the main branch
is always stable and that new features are also added contanstly without
the fear of things going wrong.

Changes:
[+] Now the theme uses the JetBrains Mono font for code sections.
    In my opnion, this is better.
[+] Made the placement of the "jump to present" and "mark as read"
    better to see. Now they also don't overlap other elements.
[/] Changed roundness of buttons, to make them more rounded, as the
    old way made the buttons inconsistent.
[/] Fixed disabled text field not hiding itself.
[/] Added shadow to the bar that shows the server you're in. Making it
    more intuitive that it overlaps other elements.
[/] Made the pop-up lists be a little bit distanced from the selector.