Commit focused on groups
This commit currently has a problem I'm already aware of, which is the
fact that the members list placement when you are in a group call is
wrong. I have tried to fix this but to no avail. This was also affecting
servers so since I wanted to iron out everything server-related I added
this unfinished change.
Aside from that, not much has been changed. Mainly all of the changes
are related to removing redundant code.