Font size update on code snippets
For some unknown reason I changed the code snippet's font size without testing it properly so a lot of elements in Discord were with the wrong size.
This update just removes this line, for now it's commented since I don't exactly remember the reason I added it, thus it might be useful.
In the future I will either remove it entirely or make it useful and not break elements.
Sorry