White theme Support (WIP)
This theme adds the long needed white theme support. Right now it works
but some elements need better colors to fit more with the whole thing
better, hence why it's an WIP. This update comes with a better color
management and (possibly) a future customization of some colors!

Additions:
 - White theme support (WIP)

Fixes:
 - Proper padding of call region
 - Animation fixes
 - Addition of some elements (mostly ads) to the "not display" list
 - Name of the variable of the current user/server you're in. Shouldn't
   affect anything

Bugs:
 - Some elements may not have the proper color palette in white mode, as
   some colors are hardcoded for now