Emergency commit
This commit is here because I think the scoped CSS library Discord uses
got a update so all elements that worked previously don't work now.

A lot of stuff are still broken but this will be fixed within the next
days.

A blog post will be posted about this after all the fixes are done.

Fixes:
 - (Sort of) made the theme work again