Second emergency commit
This commit fixes A LOT of elements that were previously not working
with the last commit. I won't cover all the changes here since it's a
lot of stuff. If you are interested. Please scroll through all the
changes and you will understand why I'm not listing them.

Also, now in the README it tells you to go the official website for the
installation steps. As said previously, after everything is fixed, I
will make a blogpost sharing my thoughts on this chaotic update.